[fix] Requested changes moved caching mixins to separate file
